adjust padding. update  \setcvskilllegendcolumns
Finetune paddings in legend. Some paddings in cvbodyiv are by eyenorm.

Move calculation of desriptorwidths to the \cvskilllengend and
\cvskillplainlegend commands to recalculate these lengths on the
fly. This way, the recalculation of those width is closer to where the
legnths get used which makes the code more readable and makes it easier
to adjust padding.
In addition, this step allows dropping recalculation of these lengths from the
\setcvskilllegendcolumns method, cleaning this method up and achieving a cleaner code.