remove pdfs completely from repo
due to the new Makefile the pdf files have become obsolete in the repo.
The new release rule of the make file creates all templates and the user
guide and includes them into the tarball.
In this way even users for which the linux based Makefile does not work
can download precompiled examples and and the manual